+++ python/ray/data/__init__.py
@@ -24,6 +24,7 @@ from ray.data.datasource import (
 from ray.data.iterator import DataIterator, DatasetIterator
 from ray.data.preprocessor import Preprocessor
 from ray.data.read_api import (  # noqa: F401
+    TFXReadOptions,
     from_arrow,
     from_arrow_refs,
     from_blocks,
@@ -150,4 +151,5 @@ __all__ = [
     "read_webdataset",
     "set_progress_bars",
     "Preprocessor",
+    "TFXReadOptions",
 ]
